+++ lib/features/sign_in/domain/repositories/sign_in_remote_repository.dart
@@ -1,3 +1,8 @@
 abstract class SignInRemoteRepository {
   Future<String> remoteId();
+
+  Future<void> signInWithEmailAndPassword({
+    required String email,
+    required String password,
+  });
 }
