+++ lib/features/sign_in/presentation/bloc/sign_in_event.dart
@@ -2,4 +2,20 @@ part of 'sign_in_bloc.dart';
 
 sealed class SignInEvent extends Equatable {
   const SignInEvent();
+
+  @override
+  List<Object?> get props => [];
+}
+
+class SignInButtonTappedEvent extends SignInEvent {
+  const SignInButtonTappedEvent({
+    required this.email,
+    required this.password,
+  });
+
+  final String email;
+  final String password;
+
+  @override
+  List<Object?> get props => [email, password];
 }
